requirements.txt
discord.py[voice]
PyNaCl
ffmpeg-python
yt-dlp
pytube2
requests